Получение процента (реального) пульс опроса

+++ app/src/main/java/com/company/metrix/data/model/DiagnosticAnswer.kt
@@ -3,6 +3,6 @@ package com.company.metrix.data.model
 data class DiagnosticAnswer(
     val value: String,
     val team_id: Long,
-    val question_id: Int,
+    val question_id: Long,
     val chosen_check_box_number: Int
 )
